backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.0
motor==3.6.0
# Pin PyMongo to a version compatible with Motor 3.6.x
pymongo==4.9.2
redis==5.0.4
loguru==0.7.2
pytest==8.2.2
pytest-asyncio==0.23.7
anyio==4.4.0
anthropic==0.39.0
# Use numpy <2.0 because langchain 0.2.x requires it on Python >=3.12
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
lightgbm==4.5.0
scipy==1.16.1
optuna==3.6.1
shap==0.44.1
beautifulsoup4==4.12.3
lxml==5.2.2
scrapy==2.11.2
selenium==4.23.1
google-cloud-bigquery==3.25.0
langchain==0.2.10
langchain-anthropic==0.1.19
langgraph==0.2.1


